Outputs/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
PyJWT==2.10.1
pydantic==2.10.6
requests==2.32.3
pytest==8.3.4
httpx==0.28.1
